requirements.txt
distlib==0.3.6
filelock==3.12.0
mysql-connector-python==8.0.33
numpy==1.24.3
opencv-python==4.7.0.72
platformdirs==3.5.1
pytz==2023.3
Unidecode==1.3.6
virtualenv==20.23.0
